优化高度相关的代理方法，规定返回整型

+++ JXPagerViewExample-OC/JXPagerViewExample-OC/Example/Common/PagingViewController.m
@@ -78,11 +78,11 @@
     return self.userHeaderView;
 }
 
-- (CGFloat)tableHeaderViewHeightInPagerView:(JXPagerView *)pagerView {
+- (NSUInteger)tableHeaderViewHeightInPagerView:(JXPagerView *)pagerView {
     return JXTableHeaderViewHeight;
 }
 
-- (CGFloat)heightForPinSectionHeaderInPagerView:(JXPagerView *)pagerView {
+- (NSUInteger)heightForPinSectionHeaderInPagerView:(JXPagerView *)pagerView {
     return JXheightForHeaderInSection;
 }
