+++ JXPagingView/Example/CollectionView/CollectionViewViewController.swift
@@ -31,11 +31,11 @@ class CollectionViewViewController: UIViewController {
 
         listViewArray = [powerListView, hobbyListView, partnerListView]
 
-        userHeaderContainerView = UIView(frame: CGRect(x: 0, y: 0, width: UIScreen.main.bounds.size.width, height: JXTableHeaderViewHeight))
+        userHeaderContainerView = UIView(frame: CGRect(x: 0, y: 0, width: UIScreen.main.bounds.size.width, height: CGFloat(JXTableHeaderViewHeight)))
         userHeaderView = PagingViewTableHeaderView(frame: userHeaderContainerView.bounds)
         userHeaderContainerView.addSubview(userHeaderView)
 
-        categoryView = JXCategoryTitleView(frame: CGRect(x: 0, y: 0, width: UIScreen.main.bounds.size.width, height: JXheightForHeaderInSection))
+        categoryView = JXCategoryTitleView(frame: CGRect(x: 0, y: 0, width: UIScreen.main.bounds.size.width, height: CGFloat(JXheightForHeaderInSection)))
         categoryView.titles = titles
         categoryView.backgroundColor = UIColor.white
         categoryView.titleSelectedColor = UIColor(red: 105/255, green: 144/255, blue: 239/255, alpha: 1)
@@ -77,7 +77,7 @@ class CollectionViewViewController: UIViewController {
 
 extension CollectionViewViewController: JXPagingViewDelegate {
 
-    func tableHeaderViewHeight(in pagingView: JXPagingView) -> CGFloat {
+    func tableHeaderViewHeight(in pagingView: JXPagingView) -> Int {
         return JXTableHeaderViewHeight
     }
 
@@ -85,7 +85,7 @@ extension CollectionViewViewController: JXPagingViewDelegate {
         return userHeaderContainerView
     }
 
-    func heightForPinSectionHeader(in pagingView: JXPagingView) -> CGFloat {
+    func heightForPinSectionHeader(in pagingView: JXPagingView) -> Int {
         return JXheightForHeaderInSection
     }
